requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pillow==10.4.0
mlx
mlx-vlm==0.1.20
transformers==4.49.0
torch==2.3.1
torchvision==0.18.1
